# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Apartment Management API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/apartment-areas/{id}": {
            "get": {
                "tags": [
                    "ApartmentAreas"
                ],
                "summary": "Get a detail apartment-areas",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Display a apartment areas."
                    }
                }
            }
        },
        "/api/apartment-areas/create": {
            "post": {
                "tags": [
                    "ApartmentAreas"
                ],
                "summary": "Add a new apartment areas",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "floors": {
                                        "type": "integer"
                                    },
                                    "address": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Add a new apartment areas"
                    }
                }
            }
        },
        "/api/apartment-areas/{id}/update": {
            "put": {
                "tags": [
                    "ApartmentAreas"
                ],
                "summary": "Update a apartment-areas",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path"
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "floors": {
                                        "type": "integer"
                                    },
                                    "address": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Update a apartment areas"
                    }
                }
            }
        },
        "/api/apartment-areas/{id}/delete": {
            "delete": {
                "tags": [
                    "ApartmentAreas"
                ],
                "summary": "Delete a apartment-areas",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete a apartment areas."
                    }
                }
            }
        },
        "/api/apartment": {
            "get": {
                "tags": [
                    "Apartment"
                ],
                "summary": "Get all apartments",
                "parameters": [
                    {
                        "name": "apartment_areas",
                        "in": "path"
                    },
                    {
                        "name": "sortBy",
                        "in": "path"
                    },
                    {
                        "name": "sortOrder",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Display a listing of apartment."
                    }
                }
            }
        },
        "/api/apartment/{id}": {
            "get": {
                "tags": [
                    "Apartment"
                ],
                "summary": "Get a detail apartment",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Display a apartment."
                    }
                }
            }
        },
        "/api/apartment/create": {
            "post": {
                "tags": [
                    "Apartment"
                ],
                "summary": "Add a new apartment",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    },
                                    "area": {
                                        "type": "float"
                                    },
                                    "price": {
                                        "type": "float"
                                    },
                                    "rooms": {
                                        "type": "integer"
                                    },
                                    "apartment_areas_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Add a new apartment"
                    }
                }
            }
        },
        "/api/apartment/{id}/update": {
            "put": {
                "tags": [
                    "Apartment"
                ],
                "summary": "Update a apartment",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path"
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "area": {
                                        "type": "float"
                                    },
                                    "price": {
                                        "type": "float"
                                    },
                                    "rooms": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Update a apartment"
                    }
                }
            }
        },
        "/api/apartment/{id}/delete": {
            "delete": {
                "tags": [
                    "Apartment"
                ],
                "summary": "Delete a apartment",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Display a apartment."
                    }
                }
            }
        },
...
    },
    "security": [
        []
    ]
}